pyproject.toml
[tool.poetry]
name = "assistent"
version = "0.1.0"
description = "Dain Assistent"
authors = ["Silvio Jurk"]

[tool.poetry.dependencies]
setuptools = "^62.1.0"
python = ">=3.8,<3.11"
numpy = "^1.21.5"
pandas = "^1.3.5"
tqdm = "^4.64.0"
jsons = "^1.6.3"
azure-cognitiveservices-speech = "1.27.0"
openai = "^0.27.4"
python-dotenv = "^1.0.0"
pyttsx3 = "^2.9.0"
SpeechRecognition = "^3.10.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^22.3.0"
isort = "^5.10.1"
pytest-watch = "^4.2.0"
mypy = "^0.942"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"